Made changes to cpudriver and system/hardware_info. Added a shared header which needs fixing since I did not make the connections in my other headers that include it correctly
