+++ System_Info/System_Info.vcxproj
@@ -143,7 +143,7 @@
     <ClInclude Include="processor.h" />
     <ClInclude Include="projutils.h" />
     <ClInclude Include="storagedevice.h" />
-  </ItemGroup>
+    <ClInclude Include=".\..\shared_headers\cpu_shared_info.h">
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
